requirements.txt
boto3==1.4.8
botocore==1.8.2
certifi==2017.11.5
docutils==0.14
Flask==0.12.2
gspread==0.6.2
Jinja2==2.10
MarkupSafe==1.0
mypy==0.550
oauth2client==4.1.2
psutil==5.4.1
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
